web/hitparade/requirements.txt
beautifulsoup4==4.6.0
boto3==1.4.4
botocore==1.5.45
coverage==4.4
Django==1.10.3
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-dynamic-fixture==1.9.5
django-extensions==1.7.7
django-model-utils==3.0.0
django-mysql==1.1.1
django-nose==1.4.4
django-user-accounts==2.0.0
django-webtest==1.9.1
djangorestframework==3.6.2
docutils==0.13.1
funcsigs==1.0.2
futures==3.1.1
jmespath==0.9.2
jsonfield==2.0.0
mock==2.0.0
MySQL-python==1.2.5
nose==1.3.7
oauthlib==2.0.1
pbr==3.0.0
pinax-eventlog==1.1.2
pinax-theme-bootstrap==7.11.0
pinax-webanalytics==2.0.4
py==1.4.33
PyJWT==1.4.2
pytest==3.0.7
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.8.0
s3transfer==0.1.10
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
sure==1.4.6
uWSGI==2.0.14
virtualenv==15.1.0
waitress==1.0.2
WebOb==1.7.2
WebTest==2.0.27
git+https://github.com/stattleship/stattleship-python
